Moved Punch to components folder
